Fix persistence on tamed pets. Fixes BUKKIT-3300
With the persistence api introduced, pets did not have their
persistence flag updated to reflect their persistence. This caused
tame wolves and ocelots to not persist under specific conditions.

This commit fixes that by forcing all previous wolves and ocelots to
be tame, and persists pets automatically when tamed. Allowing plugins
to override it as well if necessary.